add estimated weight to shutttle service item on MTO page, move propTypes to shape
