+++ src/components/Office/RequestedServiceItemsTable/RequestedServiceItemsTable.jsx
@@ -1,8 +1,8 @@
 import React from 'react';
 import PropTypes from 'prop-types';
 
-import { MTOServiceItemDimensionShape, MTOServiceItemCustomerContactShape } from '../../../types/order';
 import { SERVICE_ITEM_STATUS } from '../../../shared/constants';
+import { ServiceItemDetailsShape } from '../../../types/serviceItems';
 
 import styles from './RequestedServiceItemsTable.module.scss';
 
@@ -51,25 +51,7 @@ RequestedServiceItemsTable.propTypes = {
   handleUpdateMTOServiceItemStatus: PropTypes.func.isRequired,
   handleShowRejectionDialog: PropTypes.func.isRequired,
   statusForTableType: PropTypes.string.isRequired,
-  serviceItems: PropTypes.arrayOf(
-    PropTypes.shape({
-      id: PropTypes.string,
-      mtoShipmentID: PropTypes.string,
-      createdAt: PropTypes.string,
-      approvedAt: PropTypes.string,
-      rejectedAt: PropTypes.string,
-      serviceItem: PropTypes.string,
-      code: PropTypes.string,
-      details: PropTypes.shape({
-        reason: PropTypes.string,
-        pickupPostalCode: PropTypes.string,
-        itemDimensions: MTOServiceItemDimensionShape,
-        crateDimensions: MTOServiceItemDimensionShape,
-        firstCustContact: MTOServiceItemCustomerContactShape,
-        secondCustContact: MTOServiceItemCustomerContactShape,
-      }),
-    }),
-  ).isRequired,
+  serviceItems: PropTypes.arrayOf(ServiceItemDetailsShape).isRequired,
 };
 
 export default RequestedServiceItemsTable;
