+++ src/components/Office/ServiceItemDetails/ServiceItemDetails.jsx
@@ -1,10 +1,11 @@
 import React from 'react';
-import { PropTypes } from 'prop-types';
+
+import { ServiceItemDetailsShape } from '../../../types/serviceItems';
 
 import styles from './ServiceItemDetails.module.scss';
 
 import { formatDate } from 'shared/dates';
-import { convertFromThousandthInchToInch } from 'shared/formatters';
+import { convertFromThousandthInchToInch, formatWeight } from 'shared/formatters';
 
 function generateDetailText(details, id) {
   const detailList = Object.keys(details).map((detail) => (
@@ -16,7 +17,7 @@ function generateDetailText(details, id) {
   return detailList;
 }
 
-const ServiceItemDetails = ({ className, id, code, details }) => {
+const ServiceItemDetails = ({ id, code, details }) => {
   let detailSection;
   switch (code) {
     case 'DOFSIT':
@@ -90,9 +91,15 @@ const ServiceItemDetails = ({ className, id, code, details }) => {
     }
     case 'DOSHUT':
     case 'DDSHUT': {
+      const estimatedWeight = details.estimatedWeight ? formatWeight(details.estimatedWeight) : `— lbs`;
       detailSection = (
         <div>
-          <dl>{generateDetailText({ 'Estimated Weight': '', Reason: details.reason })}</dl>
+          <dl>
+            <div key={`${id}-estimatedWeight`} className={styles.detailLine}>
+              <dd className={styles.detailType}>{estimatedWeight}</dd> <dt>estimated weight</dt>
+            </div>
+            {generateDetailText({ Reason: details.reason })}
+          </dl>
         </div>
       );
       break;
@@ -100,33 +107,12 @@ const ServiceItemDetails = ({ className, id, code, details }) => {
     default:
       detailSection = <div>—</div>;
   }
-  return <div className={className}>{detailSection}</div>;
+  return <div>{detailSection}</div>;
 };
 
-ServiceItemDetails.propTypes = {
-  className: PropTypes.string,
-  id: PropTypes.string.isRequired,
-  code: PropTypes.string.isRequired,
-  details: PropTypes.shape({
-    description: PropTypes.string,
-    pickupPostalCode: PropTypes.string,
-    reason: PropTypes.string,
-    itemDimensions: PropTypes.shape({ length: PropTypes.number, width: PropTypes.number, height: PropTypes.number }),
-    crateDimensions: PropTypes.shape({ length: PropTypes.number, width: PropTypes.number, height: PropTypes.number }),
-    firstCustomerContact: PropTypes.shape({
-      timeMilitary: PropTypes.string,
-      firstAvailableDeliveryDate: PropTypes.string,
-    }),
-    secondCustomerContact: PropTypes.shape({
-      timeMilitary: PropTypes.string,
-      firstAvailableDeliveryDate: PropTypes.string,
-    }),
-  }),
-};
+ServiceItemDetails.propTypes = ServiceItemDetailsShape.isRequired;
 
 ServiceItemDetails.defaultProps = {
-  className: undefined,
   details: {},
 };
-
 export default ServiceItemDetails;
